+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# What's New
+## Version 0.0.7
+
+1. Fixed tasklist error that not refreshing the skin.
+2. Added Rainmeter Logs Webview Panel. Enable now from the statusbar.
 
 ## Version 0.0.6